+++ include/metashell/is_template_type.hpp
@@ -17,11 +17,11 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <string>
+#include <metashell/data/type.hpp>
 
 namespace metashell {
 
-bool is_template_type(const std::string& type);
+bool is_template_type(const data::type& type);
 
 }