+++ lib/core/metaprogram.cpp
@@ -35,12 +35,12 @@ metaprogram::metaprogram(
   mode(mode),
   evaluation_result(evaluation_result)
 {
-  root_vertex = add_vertex(root_name, root_source_location);
+  root_vertex = add_vertex(data::type(root_name), root_source_location);
   reset_state();
 }
 
 metaprogram::vertex_descriptor metaprogram::add_vertex(
-  const std::string& name,
+  const data::type& type,
   const data::file_location& source_location)
 {
   vertex_descriptor vertex = boost::add_vertex(graph);
@@ -53,7 +53,7 @@ metaprogram::vertex_descriptor metaprogram::add_vertex(
 
   auto& vertex_property = get_vertex_property(vertex);
 
-  vertex_property.name = name;
+  vertex_property.type = type;
   vertex_property.source_location = source_location;
 
   return vertex;
@@ -380,17 +380,15 @@ metaprogram::optional_edge_descriptor metaprogram::get_current_edge() const {
 data::frame metaprogram::to_frame(const edge_descriptor& e_) const
 {
   const auto& vp = get_vertex_property(get_target(e_));
-  const data::type t(vp.name);
-
   const auto& ep = get_edge_property(e_);
 
   switch (get_mode()) {
     case mode_t::normal:
       return data::frame(
-        t, vp.source_location, ep.point_of_instantiation, ep.kind);
+        vp.type, vp.source_location, ep.point_of_instantiation, ep.kind);
     default:
     case mode_t::full:
-      return data::frame(t, vp.source_location);
+      return data::frame(vp.type, vp.source_location);
     case mode_t::profile:
       double ratio = [&] {
         if (full_time_taken <= 0.0) {
@@ -400,7 +398,7 @@ data::frame metaprogram::to_frame(const edge_descriptor& e_) const
         }
       }();
       return data::frame(
-        t, vp.source_location, ep.point_of_instantiation, ep.kind,
+        vp.type, vp.source_location, ep.point_of_instantiation, ep.kind,
         ep.time_taken, ratio);
   };
 }
@@ -414,7 +412,7 @@ data::frame metaprogram::get_current_frame() const {
 
 data::frame metaprogram::get_root_frame() const {
   const auto& vp = get_vertex_property(get_root_vertex());
-  return data::frame(data::type(vp.name), vp.source_location);
+  return data::frame(vp.type, vp.source_location);
 }
 
 data::backtrace metaprogram::get_backtrace() const {