+++ test/system/app/mdb/test_continue.cpp
@@ -103,7 +103,7 @@ JUST_TEST_CASE(test_mdb_continue_fibonacci_1_breakpoint) {
 
   auto i = r.begin() + 6;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<0>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<0>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<0>"), _, _, instantiation_kind::memoization), *i);
 }
@@ -121,7 +121,7 @@ JUST_TEST_CASE(test_mdb_continue_2_fibonacci_1_breakpoint) {
 
   auto i = r.begin() + 6;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::memoization), *i);
 }
@@ -140,12 +140,12 @@ JUST_TEST_CASE(test_mdb_continue_twice_fibonacci_1_breakpoint) {
 
   auto i = r.begin() + 6;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::template_instantiation), *i);
 
   i += 2;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::memoization), *i);
 }
@@ -165,12 +165,12 @@ JUST_TEST_CASE(test_mdb_continue_fibonacci_2_breakpoints) {
 
   auto i = r.begin() + 8;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<6>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 2: regex(\"fib<6>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<6>"), _, _, instantiation_kind::template_instantiation), *i);
 
   i += 2;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::template_instantiation), *i);
 }
@@ -189,7 +189,7 @@ JUST_TEST_CASE(test_mdb_continue_2_fibonacci_2_breakpoints) {
 
   auto i = r.begin() + 8;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::template_instantiation), *i);
 }
@@ -270,13 +270,13 @@ JUST_TEST_CASE(test_mdb_continue_minus_1_with_preceding_breakpoint) {
     );
   auto i = r.begin() + 8;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 2: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::template_instantiation), *i);
 
   i += 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<6>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<6>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<6>"), _, _, instantiation_kind::template_instantiation), *i);
 }
@@ -294,7 +294,7 @@ JUST_TEST_CASE(test_mdb_continue_minus_1_without_preceding_breakpoint) {
     );
   auto i = r.begin() + 6;
 
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL(
       frame(type("fib<5>"), _, _, instantiation_kind::template_instantiation), *i);
 
@@ -366,7 +366,7 @@ JUST_TEST_CASE(
       *i);
 
   i += 2;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"fib<5>\" reached"), *i++);
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"fib<5>\") reached"), *i++);
   JUST_ASSERT_EQUAL( frame(type("fib<5>")), *i);
 
   i += 2;