+++ test/system/app/mdb/test_mdb_evaluate.cpp
@@ -41,7 +41,7 @@ JUST_TEST_CASE(test_mdb_evaluate_int) {
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   ++i;
   JUST_ASSERT_EQUAL(type("int"), *i);
 }
@@ -59,7 +59,7 @@ JUST_TEST_CASE(test_mdb_evaluate_fib_10) {
 
   auto i = r.begin() + 3;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   ++i;
   JUST_ASSERT_EQUAL(type("int_<55>"), *i);
 }
@@ -76,7 +76,7 @@ JUST_TEST_CASE(test_mdb_evaluate_empty_environment) {
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   ++i;
   JUST_ASSERT_EQUAL(prompt("(mdb)"), *i);
 }
@@ -124,8 +124,8 @@ JUST_TEST_CASE(test_mdb_evaluate_failure_will_reset_metaprogram_state) {
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
-  JUST_ASSERT_EQUAL(error(_), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
+  JUST_ASSERT_EQUAL(error(_), *i); i += 2;
   JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i);
 }
 
@@ -158,8 +158,8 @@ JUST_TEST_CASE(test_mdb_evaluate_missing_argument_will_run_last_metaprogram) {
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 3;
   JUST_ASSERT_EQUAL(type("int"), *i);
 }
 
@@ -200,10 +200,10 @@ JUST_TEST_CASE(test_mdb_reevaluate_environment) {
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   ++i;
   JUST_ASSERT_EQUAL(prompt("(mdb)"), *i); ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   ++i;
   JUST_ASSERT_EQUAL(prompt("(mdb)"), *i);
 }
@@ -221,7 +221,7 @@ JUST_TEST_CASE(test_mdb_evaluate_filters_similar_edges) {
 
   auto i = r.begin() + 3;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   JUST_ASSERT_EQUAL(
     call_graph(
       {
@@ -241,83 +241,80 @@ JUST_TEST_CASE(test_mdb_evaluate_filters_similar_edges) {
 }
 
 JUST_TEST_CASE(test_mdb_evaluate_clears_breakpoints) {
-  // TODO: add mdb command to list breakpoints
   const auto r =
     run_metashell(
       {
         command("#msh mdb"),
         command("evaluate int"),
         command("rbreak int"),
-        command("c"),
+        command("break list"),
         command("evaluate unsigned int"),
-        command("c")
+        command("break list"),
       }
     );
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   JUST_ASSERT_EQUAL(
     raw_text("Breakpoint \"int\" will stop the execution on 1 location"),
     *i
   );
-  ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"int\" reached"), *i); ++i; ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram finished"), *i);
+  i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"int\")"), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("No breakpoints currently set"), *i);
 }
 
 JUST_TEST_CASE(test_mdb_evaluate_reevaluate_clears_breakpoints) {
-  // TODO: add mdb command to list breakpoints
   const auto r =
     run_metashell(
       {
         command("#msh mdb"),
         command("evaluate int"),
         command("rbreak int"),
-        command("c"),
+        command("break list"),
         command("evaluate"),
-        command("c")
+        command("break list")
       }
     );
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   JUST_ASSERT_EQUAL(
     raw_text("Breakpoint \"int\" will stop the execution on 1 location"),
     *i
   );
-  ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"int\" reached"), *i); ++i; ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram finished"), *i);
+  i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"int\")"), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("No breakpoints currently set"), *i);
 }
 
 JUST_TEST_CASE(test_mdb_evaluate_failure_clears_breakpoints) {
-  // TODO: add mdb command to list breakpoints
   const auto r =
     run_metashell(
       {
         command("#msh mdb"),
         command("evaluate int"),
         command("rbreak int"),
-        command("c"),
+        command("break list"),
         command("evaluate asd"),
-        command("c")
+        command("break list")
       }
     );
 
   auto i = r.begin() + 2;
 
-  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); ++i; ++i;
+  JUST_ASSERT_EQUAL(raw_text("Metaprogram started"), *i); i += 2;
   JUST_ASSERT_EQUAL(
     raw_text("Breakpoint \"int\" will stop the execution on 1 location"),
     *i
   );
-  ++i; ++i;
-  JUST_ASSERT_EQUAL(raw_text("Breakpoint \"int\" reached"), *i); ++i; ++i; ++i;
-  JUST_ASSERT_EQUAL(error(_), *i); ++i; ++i;
-  JUST_ASSERT_EQUAL(error("Metaprogram not evaluated yet"), *i);
+  i += 2;
+  JUST_ASSERT_EQUAL(raw_text("Breakpoint 1: regex(\"int\")"), *i); i += 2;
+  JUST_ASSERT_EQUAL(error(_), *i); i += 2;
+  JUST_ASSERT_EQUAL(raw_text("No breakpoints currently set"), *i);
 }