+++ test/unit/test_metaprogram.cpp
@@ -63,8 +63,8 @@ JUST_TEST_CASE(test_metaprogram_constructor) {
   JUST_ASSERT_EQUAL(mp.get_num_edges(), 0u);
 
   JUST_ASSERT_EQUAL(
-      mp.get_vertex_property(mp.get_root_vertex()).name,
-      "some_type");
+      mp.get_vertex_property(mp.get_root_vertex()).type,
+      data::type("some_type"));
 
   assert_state_equal(mp.get_state(),
         {false},
@@ -94,16 +94,17 @@ JUST_TEST_CASE(test_metaprogram_with_single_non_root_vertex) {
 
   JUST_ASSERT_EQUAL(mp.get_mode(), metaprogram::mode_t::normal);
 
-  auto vertex_a = mp.add_vertex("A", data::file_location("a.hpp", 10, 20));
+  auto vertex_a = mp.add_vertex(
+      data::type("A"), data::file_location("a.hpp", 10, 20));
   auto edge_root_a = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::template_instantiation,
-        data::file_location("foo.cpp", 10, 20),
-        10.0);
+      data::instantiation_kind::template_instantiation,
+      data::file_location("foo.cpp", 10, 20),
+      10.0);
 
   JUST_ASSERT_EQUAL(mp.get_num_vertices(), 2u);
   JUST_ASSERT_EQUAL(mp.get_num_edges(), 1u);
 
-  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).name, "A");
+  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).type.name(), "A");
   JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).source_location,
       data::file_location("a.hpp", 10, 20));
 
@@ -150,20 +151,21 @@ JUST_TEST_CASE(test_metaprogram_with_single_non_root_vertex_parallel_edge) {
 
   JUST_ASSERT_EQUAL(mp.get_mode(), metaprogram::mode_t::normal);
 
-  auto vertex_a = mp.add_vertex("A", data::file_location("b.hpp", 40, 50));
+  auto vertex_a = mp.add_vertex(
+      data::type("A"), data::file_location("b.hpp", 40, 50));
   auto edge_root_a_ti = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::template_instantiation,
-        data::file_location("bar.cpp", 20, 10),
-        10.0);
+      data::instantiation_kind::template_instantiation,
+      data::file_location("bar.cpp", 20, 10),
+      10.0);
   auto edge_root_a_me = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::memoization,
-        data::file_location("foobar.cpp", 21, 11),
-        10.0);
+      data::instantiation_kind::memoization,
+      data::file_location("foobar.cpp", 21, 11),
+      10.0);
 
   JUST_ASSERT_EQUAL(mp.get_num_vertices(), 2u);
   JUST_ASSERT_EQUAL(mp.get_num_edges(), 2u);
 
-  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).name, "A");
+  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).type.name(), "A");
   JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).source_location,
       data::file_location("b.hpp", 40, 50));
 
@@ -222,7 +224,8 @@ JUST_TEST_CASE(test_metaprogram_step_back_with_single_non_root_vertex) {
       data::file_location{},
       data::type_or_error(data::type("the_result_type")));
 
-  auto vertex_a = mp.add_vertex("A", data::file_location("c.hpp", 30, 35));
+  auto vertex_a = mp.add_vertex(
+      data::type("A"), data::file_location("c.hpp", 30, 35));
   auto edge_root_a = mp.add_edge(mp.get_root_vertex(), vertex_a,
         data::instantiation_kind::template_instantiation,
         data::file_location("foobar.cpp", 21, 11),
@@ -231,7 +234,7 @@ JUST_TEST_CASE(test_metaprogram_step_back_with_single_non_root_vertex) {
   JUST_ASSERT_EQUAL(mp.get_num_vertices(), 2u);
   JUST_ASSERT_EQUAL(mp.get_num_edges(), 1u);
 
-  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).name, "A");
+  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).type.name(), "A");
   JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).source_location,
       data::file_location("c.hpp", 30, 35));
 
@@ -278,20 +281,21 @@ JUST_TEST_CASE(
       data::file_location{},
       data::type_or_error(data::type("the_result_type")));
 
-  auto vertex_a = mp.add_vertex("A", data::file_location("d.hpp", 10, 11));
+  auto vertex_a = mp.add_vertex(
+      data::type("A"), data::file_location("d.hpp", 10, 11));
   auto edge_root_a_ti = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::template_instantiation,
-        data::file_location("xx.cpp", 1, 2),
-        10.0);
+      data::instantiation_kind::template_instantiation,
+      data::file_location("xx.cpp", 1, 2),
+      10.0);
   auto edge_root_a_me = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::memoization,
-        data::file_location("yy.cpp", 1, 2),
-        10.0);
+      data::instantiation_kind::memoization,
+      data::file_location("yy.cpp", 1, 2),
+      10.0);
 
   JUST_ASSERT_EQUAL(mp.get_num_vertices(), 2u);
   JUST_ASSERT_EQUAL(mp.get_num_edges(), 2u);
 
-  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).name, "A");
+  JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).type.name(), "A");
   JUST_ASSERT_EQUAL(mp.get_vertex_property(vertex_a).source_location,
       data::file_location("d.hpp", 10, 11));
 
@@ -380,17 +384,19 @@ JUST_TEST_CASE(test_metaprogram_step_sorting_in_profile_mode) {
       data::file_location{},
       data::type_or_error(data::type("the_result_type")));
 
-  auto vertex_a = mp.add_vertex("A", data::file_location("a.hpp", 0, 0));
-  auto vertex_b = mp.add_vertex("B", data::file_location("b.hpp", 1, 2));
+  auto vertex_a = mp.add_vertex(
+      data::type("A"), data::file_location("a.hpp", 0, 0));
+  auto vertex_b = mp.add_vertex(
+      data::type("B"), data::file_location("b.hpp", 1, 2));
 
   auto edge_root_a_ti = mp.add_edge(mp.get_root_vertex(), vertex_a,
-        data::instantiation_kind::template_instantiation,
-        data::file_location("xx.cpp", 1, 2),
-        0.0);
+      data::instantiation_kind::template_instantiation,
+      data::file_location("xx.cpp", 1, 2),
+      0.0);
   auto edge_root_b_ti = mp.add_edge(mp.get_root_vertex(), vertex_b,
-        data::instantiation_kind::memoization,
-        data::file_location("yy.cpp", 1, 2),
-        10.0);
+      data::instantiation_kind::memoization,
+      data::file_location("yy.cpp", 1, 2),
+      10.0);
 
   mp.get_edge_property(edge_root_a_ti).time_taken = 10.0;
   mp.get_edge_property(edge_root_b_ti).time_taken = 30.0;
@@ -416,7 +422,7 @@ JUST_TEST_CASE(test_metaprogram_step_sorting_in_profile_mode) {
     JUST_ASSERT(frame.is_full());
     JUST_ASSERT(frame.is_profiled());
 
-    JUST_ASSERT_EQUAL("B", frame.name().name());
+    JUST_ASSERT_EQUAL("B", frame.type().name());
     JUST_ASSERT_EQUAL("yy.cpp", frame.point_of_instantiation().name);
     JUST_ASSERT_EQUAL(30.0, frame.time_taken());
     JUST_ASSERT_EQUAL(0.75, frame.time_taken_ratio());
@@ -432,7 +438,7 @@ JUST_TEST_CASE(test_metaprogram_step_sorting_in_profile_mode) {
 
     // A should come second, since that was faster
 
-    JUST_ASSERT_EQUAL("A", frame.name().name());
+    JUST_ASSERT_EQUAL("A", frame.type().name());
     JUST_ASSERT_EQUAL("xx.cpp", frame.point_of_instantiation().name);
     JUST_ASSERT_EQUAL(10.0, frame.time_taken());
     JUST_ASSERT_EQUAL(0.25, frame.time_taken_ratio());